+++ lib/hcloud/resources/pricing.rb
@@ -32,5 +32,13 @@ module HCloud
     attribute :vat_rate
 
     attribute :volume, :volume_price
+
+    def traffic
+      # FIXME: Attribute will return null on 2024-08-05
+      # FIXME: Attribute will be removed on 2024-11-05
+      warn "[DEPRECATION] Field \"traffic\" on pricing is deprecated."
+
+      super
+    end
   end
 end